Moved shadow tests to the SkillTest file, commented out the base class test for now.
